reduce task responsibilities to key extraction only
