+++ phing/build.xml
@@ -9,7 +9,7 @@
   <taskdef name="filterFileListByFileSet" classname="phingcludes.FilterFileListByFileSetTask"/>
   <taskdef name="verbosityTask" classname="phingcludes.VerbosityTask"/>
   <taskdef name="phpVariable" classname="phingcludes.PhpVariableTask"/>
-  <taskdef name="yamlVariable" classname="phingcludes.YamlVariableTask"/>
+  <taskdef name="yamlKeys" classname="phingcludes.YamlKeysTask"/>
 
   <if>
     <available property="xdebug.enabled" extension="xdebug"/>
